change model and api
1. UserProfile을 기획 및 FE 의견에 맞게 수정
2. profile을 만들면 user에서 profile_nid와 매칭되게끔 수정
3. 따라서 회원을 만들고 profile을 또 만들어야함..
