+++ src/main/java/com/cliff/jigsaw/common/config/QueryDslConfig.java
@@ -1,0 +1,2 @@
+package com.cliff.jigsaw.common.config;public class QueryDslConfig {
+}
