+++ src/main/java/com/cliff/jigsaw/model/user/vo/profile/CreateUserProfileVo.java
@@ -1,0 +1,9 @@
+package com.cliff.jigsaw.model.user.vo.profile;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class CreateUserProfileVo {
+}
