+++ src/main/java/com/cliff/jigsaw/model/user/vo/profile/GetUserProfileVo.java
@@ -1,0 +1,14 @@
+package com.cliff.jigsaw.model.user.vo;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class UserProfileVo {
+
+    Long userProfileNid;
+    Long userNid;
+    String file;
+
+}
